update network handling
