--- src/components/Network/Network.scss
@@ -1,16 +1,0 @@
-.network {
-  @include flex(center, center, column);
-  @include fixed(0, null, 0);
-  @include sq-dimen(100%);
-  @include typeface('Nunito Sans', 700);
-
-  background-color: var(--main-color);
-  color: var(--text-color);
-  font-size: 16px;
-  z-index: $app-region-z-index - 1;
-
-  .wifi-off-icon {
-    font-size: 60px;
-    margin-bottom: 20px;
-  }
-}
