--- src/components/Network/index.ts
@@ -1,4 +1,0 @@
-import './Network.scss';
-
-export * from './Network';
-export { Network as default } from './Network';
