+++ src/components/TaskList/TaskListHeader/index.ts
@@ -1,0 +1,4 @@
+import './TaskListHeader.scss';
+
+export * from './TaskListHeader';
+export { TaskListHeader as default } from './TaskListHeader';
